requirements.txt
requests>=2.30.0
beautifulsoup4>=4.12.0
soupsieve>=2.5.0  # beautifulsoup4的依赖
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0  # 用于密码哈希
sqlalchemy>=2.0.0  # 数据库ORM
fastapi>=0.100.0  # Web框架
uvicorn>=0.22.0  # ASGI服务器
